# Conflicts:
#	src/coreclr/ToolBox/superpmi/superpmi-shim-counter/icorjitinfo.cpp
#	src/coreclr/ToolBox/superpmi/superpmi-shim-simple/icorjitinfo.cpp
#	src/coreclr/inc/icorjitinfoimpl_generated.h
#	src/coreclr/inc/jiteeversionguid.h
#	src/coreclr/jit/ICorJitInfo_API_wrapper.hpp
#	src/coreclr/jit/instr.cpp
#	src/coreclr/jit/target.h
#	src/coreclr/tools/Common/JitInterface/CorInfoBase.cs
#	src/coreclr/tools/Common/JitInterface/ThunkGenerator/ThunkInput.txt
#	src/coreclr/tools/aot/ILCompiler.LLVM/Compiler/LLVMCodegenCompilation.cs
#	src/coreclr/tools/aot/ILCompiler.LLVM/Compiler/LLVMCodegenCompilationBuilder.cs
#	src/coreclr/tools/aot/ILCompiler.RyuJit/JitInterface/CorInfoImpl.RyuJit.cs
#	src/coreclr/tools/aot/jitinterface/jitinterface.h

Workaround for min, max, and NumItems #defines.

+++ src/coreclr/jit/jit.h
@@ -776,7 +776,7 @@ extern int jitNativeCode(CORINFO_METHOD_HANDLE methodHnd,
                          COMP_HANDLE           compHnd,
                          CORINFO_METHOD_INFO*  methodInfo,
                          void**                methodCodePtr,
-                         ULONG*                methodCodeSize,
+                         uint32_t*             methodCodeSize,
                          JitFlags*             compileFlags,
                          void*                 inlineInfoPtr);
